take StreamManager out of ConnectionManager
