+++ app/connection/index.js
@@ -1,0 +1,7 @@
+import ConnectionManager from './ConnectionManager';
+import StreamManager from './StreamManager';
+
+export {
+    ConnectionManager,
+    StreamManager,
+};
